cleaned up rendering code
